third_party/WebKit/LayoutTests/inspector/sources/debugger-frameworks/frameworks-blackbox-by-source-code.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script>
function testFunction()
{
    debugger;
    eval("239;//# sourceURL=framework.js");
}

function test()
{
    InspectorTest.startDebuggerTest(step1, true);

    function step1()
    {
        InspectorTest.addSniffer(Bindings.BlackboxManager.prototype, "_patternChangeFinishedForTests", step2);
        var frameworkRegexString = "^framework\\.js$";
        Common.settingForTest("skipStackFramesPattern").set(frameworkRegexString);
    }

    function step2()
    {
        var actions = [
            "Print", // "debugger" in testFunction()
            "StepInto", "StepInto", "Print"
        ];
        InspectorTest.waitUntilPausedAndPerformSteppingActions(actions, step4);
        InspectorTest.runTestFunction();
    }

    function step4()
    {
        InspectorTest.completeDebuggerTest();
    }
}

</script>
</head>
<body onload="runTest()">
<p>Tests provisional blackboxing.</p>
</body>
</html>
