third_party/WebKit/LayoutTests/inspector/sources/debugger-frameworks/frameworks-blackbox-by-source-code-expected.txt

Tests provisional blackboxing.

Set timer for test function.
Call stack:
    0) testFunction (frameworks-blackbox-by-source-code.html:8)

Executing StepInto...
Executing StepInto...
Call stack:
    0) testFunction (frameworks-blackbox-by-source-code.html:10)
